refactor: Move checks tftp::Bas3e to self unit-test module (u-t OK!)
Github issue #8
